added api for the JwtService, moved all files to their right packages
