+++ backend/src/main/java/com/vule/workoutcalendar/jwt/api/JwtServiceApi.java
@@ -1,0 +1,6 @@
+package com.vule.workoutcalendar.jwt.api;
+
+public interface JwtServiceApi {
+
+    Integer parseUserIdFromJwt(String jwtToken);
+}
